GPB-2 Refactor DTOs into separate package
For better organization and readability.
