+++ src/main/kotlin/de/genpare/data/dtos/LoginDTO.kt
@@ -1,4 +1,4 @@
-package de.genpare.data
+package de.genpare.data.dtos
 
 data class LoginDTO(
     val email: String
