+++ src/main/kotlin/de/genpare/data/dtos/MemberDTO.kt
@@ -1,4 +1,4 @@
-package de.genpare.data
+package de.genpare.data.dtos
 
 data class MemberDTO(
     val id: Long?,
