+++ src/main/kotlin/de/genpare/data/dtos/SessionDTO.kt
@@ -1,4 +1,4 @@
-package de.genpare.data
+package de.genpare.data.dtos
 
 data class SessionDTO(
     val sessionId: Long
